Create the api request for the signup
