+++ src/Auth/domain/repositories/Auth.repository.ts
@@ -1,7 +1,9 @@
-import { ILoginEntity, ITokenEntity } from "../models/entity";
+import { ILoginEntity, ISignupEntity, ITokenEntity } from "../models/entity";
+
 
 export interface IAuthRepository {
   // Write your functions interface
   refreshToken(refreshToken: string): Promise<ITokenEntity>;
   loginUser(loginData: ILoginEntity): Promise<ITokenEntity>;
+  signupUser(signupData: ISignupEntity): Promise<ISignupEntity>;
 }
