+++ src/Auth/infrastructure/models/adapters/index.ts
@@ -1,1 +1,3 @@
 export * from './token.adapter'
+export * from './signup.adapter'
+export * from './login.adapter'
